dashboard/requirements.txt
pandas
dash
dash_core_components
dash_html_components
plotly.graph_objs
gunicorn